Expanded readme and merged XLS_Form and XlsFormData
